Added gen_cover_img and check_pre_comp_cover
gen_cover_img will call check_pre_comp_cover to decide whether to
compile the cover image for the ebook. If the cover image exist and the
title of the cover image haven't changed then gen_cover_img won't
generate a cover image. gen_cover_img will take the norg file with the
metadata and the subtitle as its first and second arguments respectively.
The title will be extracted from the norg file.