Added write_new_modt and other bug fixes
write_new_modt will write the contents of the varible that is given as
its first argument. The varible will be formatted as "path@content".
Current one assumes that the varible only have a single item. So it won't
iterate through it. Maybe, I might need to make it iterate through each
line... And it will unsets the varible after writing to the file.

gen_cover_img used to return the path to the cover image. And to capture
this path, we needed to run it in a subshell. This has a grave
disadvantage, now it can't modify the any (modified_cover_title) global
varibles. Now it won't return anything. The path to the cover image is
calculated from the callee function itself.