Who needs to check_indexfile when you can check_metadata
Replaced check_indexfile with check_metadata. Whose stupid idea was that
to check for indexfile ? Oh that was me... hehe. Well now check_metadata
will take the path to the norg file as the first argument. And it will
check whether the file exist or not. If it exists it will check for
the "@document.meta ... @end" block. If it exists, it will return,
otherwise exits the script.