Made write_new_modt iterate through lines in variable
Now write_new_modt will iterate through each lines in the given varible
and writes the content to the specified file. This is beacause when we
generate the markdown files we will be generating markdown for the
entire directory, so markdown for any other modified norg file also will
be generated this time. Hence we need to update the lmodt_* file with
update_lmodt and write_new_modt, therefore there might be more than one
entries in the modified_time_norg varible.