Added check_pre_comp_epub and some bug fixes
Instead of simply checking for the existence of the epub,
check_pre_comp_epub will compare the lmodt of epub and the corresponding
norg file's lmodt. If the norg file(s) is compiled after epub, it will
return 1.

Now the loop which iterates through the args will find the absolute path
for the file or directory for furthur processing.