Added a mechanism to push compiled epubs to kindle
Everytime we run the script for a syllabus or note, we will add the epub
files to an array with the name of the parent directory with the forward
slashes stripped and spaces replaced with underscores. Then each of
these array names are stored in an associative array as the index called
kindle_push_files by calling add_for_push. So that epubs in the same
directory will end up in the same array. Then at the end of the script
we will call push_kindle to find the remote path for the epubs by
looking at the index.norg file in each parent directory (read the script!).
Then asks for push and pushes to the kindle using kdb (which is just
another script i wrote :)).