epubconv: epubconv will sync with the local kindle dir
previously epubconv will look for kindle and tries to push the built
epubs to the kindle. from now on epubconv will sync with the local
kindle dir so that the user only needs to simply run `kdb sync`.

if the kindle is online it will automatically runs `kdb sync` behalf of
the user, btw it won't ask for conformation